practica_3/curso-automatizador/requirements.txt
selenium==4.27.0
yt-dlp==2024.12.23
requests==2.32.3
python-dotenv==1.0.1
fake-useragent==2.0.3
Pillow>=10.0.0
img2pdf>=0.5.0
pyautogui>=0.9.54
opencv-python>=4.9.0
PyPDF2>=3.0.0
selenium-stealth==1.0.6
undetected-chromedriver==3.5.5
